requirements.txt
langchain
langchain_community
langchain_openai
fuzzywuzzy[speedup]
chromadb
pymongo
streamlit

